# speed up tsc, clean up dependencies, remove a ts-ignore