--- a/package.json
+++ b/package.json
@@ -60,8 +60,8 @@
     "relative-luminance": "2.0.1"
   },
   "peerDependencies": {
-    "react": ">=16.2.0",
-    "react-dom": ">=16.2.0",
+    "react": ">=16.8.0",
+    "react-dom": ">=16.8.0",
     "react-gemini-scrollbar": "^2.1.5 || ^2.3.0"
   },
   "devDependencies": {
@@ -86,7 +86,7 @@
     "@types/jest": "23.3.5",
     "@types/memoize-one": "3.1.1",
     "@types/node": "^10.12.26",
-    "@types/react": "16.4.16",
+    "@types/react": "16.9.23",
     "@types/react-dom": "16.0.9",
     "@types/react-tabs": "^2.3.1",
     "@types/react-test-renderer": "16.0.3",