--- a/packages/shared/components/FocusStyleManager.tsx
+++ b/packages/shared/components/FocusStyleManager.tsx
@@ -2,36 +2,30 @@ import * as React from "react";
 
 import { InteractionModeEngine } from "../../utilities/interactionMode";
 
-class FocusStyleManager extends React.PureComponent<
-  { focusEnabledClass: string; children: string | React.ReactNode },
-  {}
-> {
-  public focusWrapperRef = {};
-
-  constructor(props) {
-    super(props);
-
-    this.setRef = this.setRef.bind(this);
-  }
+class FocusStyleManager extends React.PureComponent<{
+  focusEnabledClass: string;
+  children: React.ReactElement;
+}> {
+  focusWrapperRef: Element | undefined;
 
   public componentDidMount() {
     const focusEngine = new InteractionModeEngine(
       this.focusWrapperRef as Element,
       this.props.focusEnabledClass || "focus-enabled",
-      document.documentElement as Element
+      document.documentElement
     );
     focusEngine.start();
   }
 
   public render() {
-    return React.cloneElement(React.Children.only(this.props.children), {
+    return React.cloneElement(React.Children.only(this.props.children)!, {
       ref: this.setRef
     });
   }
 
-  private setRef(ref: any) {
+  private setRef = ref => {
     this.focusWrapperRef = ref;
-  }
+  };
 }
 
 export default FocusStyleManager;