--- a/packages/shared/components/FormFieldWrapper.tsx
+++ b/packages/shared/components/FormFieldWrapper.tsx
@@ -99,9 +99,8 @@ class FormFieldWrapper extends React.PureComponent<FormFieldWrapperProps, {}> {
   private getDescribedBy(hintContent, errors) {
     if (hintContent && errors) {
       return `${hintContent} ${errors}`;
-    } else {
-      return errors || hintContent;
     }
+    return errors || hintContent;
   }
 }