--- a/packages/textInput/components/TextInputWithIcon.tsx
+++ b/packages/textInput/components/TextInputWithIcon.tsx
@@ -52,7 +52,7 @@ export class TextInputWithIcon<
   }
 
   protected getInputElementProps() {
-    let baseProps = super.getInputElementProps();
+    const baseProps = super.getInputElementProps();
     const {
       iconStart,
       iconEnd,