--- a/tsconfig.json
+++ b/tsconfig.json
@@ -11,6 +11,7 @@
     "jsx": "react",
     "noUnusedLocals": true,
     "noUnusedParameters": true,
+    "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
     "strict": true,
     "moduleResolution": "node",