1. ```
   ➜  time npx tsc
   npx tsc  16.86s user 0.80s system 168% cpu 10.460 total
   ➜  time npx tsc
   npx tsc  2.62s user 0.26s system 187% cpu 1.535 total
   #                                         ^^^^^
   ```

1. updated @react/types to a more appropriate version
2. allow for `import React from "react"` via allowSyntheticDefaultImports to
   make the tests pass
3. remove the @ts-ignore line and follow along the type errors. also fix
   some issues that come to sight then.

--------------

then my PR became red due to tslint. now there's a commit introducing sensible default (what we have in kommander and dcos-ui but without stuff that requires a lot of fixing here), running prettier on all files (formerly not everything was prettified because of "tslint-config-prettier" not being the last in the extend-block and somehow not everything being tested by the jest-test (this also is the reason why so many classes changed).